+++ CHANGELOG.md
@@ -9,6 +9,13 @@
 - Added welcome screen notes
 - Automatically adds `kotlin-metadata-jvm` to app module when using hilt and kotlin 2.3.0+
 
+### Changed
+
+- FlowStateHolder has a composable remember function with stable state class
+- ScreenStateHolder has a composable remember function with stable state class
+- Fixed convention plugin IDs added to version catalog (capitalization issue)
+- Updated version catalog handling to prevent mess (wip/in-testing)
+
 ## [0.0.4-alpha] - 2025-12-31
 
 ### Added